requirements.txt
streamlit
pinecone-client
sentence-transformers